dbt-on-astro/requirements.txt
astronomer-cosmos==1.6.0
